# feat(core, config): Simplification of Dataset Setup via config