--- a/conf/histogram-dev-source.conf
+++ b/conf/histogram-dev-source.conf
@@ -1,4 +1,21 @@
     dataset = "histogram"
+
+    definition {
+      partition-columns = ["tags:map"]
+      data-columns = ["timestamp:ts", "count:long", "sum:long", "h:hist:counter=false" ]
+      row-key-columns = [ "timestamp" ]
+      downsamplers = [ ]
+    }
+
+    options {
+      shardKeyColumns = [ "__name__", "_ns" ]
+      ignoreShardKeyColumnSuffixes = { "__name__" = [ ] }
+      valueColumn = "h"
+      metricColumn = "__name__"
+      ignoreTagsOnPartitionKeyHash = [ ]
+      copyTags = { }
+    }
+
     num-shards = 4
     min-num-nodes = 2
     sourcefactory = "filodb.kafka.KafkaIngestionStreamFactory"