--- a/conf/timeseries-dev-source.conf
+++ b/conf/timeseries-dev-source.conf
@@ -1,5 +1,27 @@
     dataset = "prometheus"
+
+    # Should not change once dataset has been set up on the server and data has been persisted to cassandra
+    definition {
+      partition-columns = ["tags:map"]
+      data-columns = ["timestamp:ts", "value:double"]
+      row-key-columns = [ "timestamp" ]
+      downsamplers = [ "tTime(0)", "dMin(1)", "dMax(1)", "dSum(1)", "dCount(1)"]
+    }
+
+    # Should not change once dataset has been set up on the server and data has been persisted to cassandra
+    options {
+      shardKeyColumns = [ "__name__", "_ns" ]
+      ignoreShardKeyColumnSuffixes = { "__name__" = ["_bucket", "_count", "_sum"] }
+      valueColumn = "value"
+      metricColumn = "__name__"
+      ignoreTagsOnPartitionKeyHash = [ "le" ]
+      copyTags = { }
+    }
+
+    # Should not change once dataset has been set up on the server and data has been persisted to cassandra
     num-shards = 4
+
+
     min-num-nodes = 2
     # Length of chunks to be written, roughly
     sourcefactory = "filodb.kafka.KafkaIngestionStreamFactory"