--- a/conf/timeseries-dev-source.conf
+++ b/conf/timeseries-dev-source.conf
@@ -1,20 +1,38 @@
     dataset = "prometheus"
 
-    # Should not change once dataset has been set up on the server and data has been persisted to cassandra
+    # Schema used for defining the BinaryRecord used in ingestion and persistence.
+    #
+    # Should not change once dataset has been set up on the server and data has been ingested into kafka
+    # or written to cassandra
     definition {
+      # defines the unique identifier for partition
       partition-columns = ["tags:map"]
+      # Schema of all of the values stored against the partition key. This includes the row-keys as well
       data-columns = ["timestamp:ts", "value:double"]
+      # Clustering key for each row. Together with partition key, they form the primary key.
       row-key-columns = [ "timestamp" ]
+      # List of downsamplers for the data columns
       downsamplers = [ "tTime(0)", "dMin(1)", "dMax(1)", "dSum(1)", "dCount(1)"]
     }
 
-    # Should not change once dataset has been set up on the server and data has been persisted to cassandra
+    # Dataset Options
+    #
+    # Should not change once dataset has been set up on the server and data has been ingested into kafka
+    # or written to cassandra
     options {
+      # These column values are used to identify the shard group for the partition
       shardKeyColumns = [ "__name__", "_ns" ]
+      # suffixes from map values are stripped before hashing
       ignoreShardKeyColumnSuffixes = { "__name__" = ["_bucket", "_count", "_sum"] }
+      # default data column name to be used referencing the value column
       valueColumn = "value"
+      # column name to use when referencing the name column
       metricColumn = "__name__"
+      # these columns are ignored in calculation of full partition key hash
       ignoreTagsOnPartitionKeyHash = [ "le" ]
+      
+      # These key-names will be replaced in the key map during ingestion/query.
+      # Temporary workaround. Will be deprecated.
       copyTags = { }
     }