--- a/conf/timeseries-dev-source.conf
+++ b/conf/timeseries-dev-source.conf
@@ -30,7 +30,7 @@
       metricColumn = "__name__"
       # these columns are ignored in calculation of full partition key hash
       ignoreTagsOnPartitionKeyHash = [ "le" ]
-      
+
       # These key-names will be replaced in the key map during ingestion/query.
       # Temporary workaround. Will be deprecated.
       copyTags = { }