--- a/conf/timeseries-dev-source.conf
+++ b/conf/timeseries-dev-source.conf
@@ -12,7 +12,7 @@
       # Clustering key for each row. Together with partition key, they form the primary key.
       row-key-columns = [ "timestamp" ]
       # List of downsamplers for the data columns
-      downsamplers = [ "tTime(0)", "dMin(1)", "dMax(1)", "dSum(1)", "dCount(1)"]
+      downsamplers = [ "tTime(0)", "dMin(1)", "dMax(1)", "dSum(1)", "dCount(1)", "dAvg(1)" ]
     }
 
     # Dataset Options