--- a/conf/timeseries-ds-1m-dev-source.conf
+++ b/conf/timeseries-ds-1m-dev-source.conf
@@ -1,4 +1,21 @@
     dataset = "prometheus_ds_1m"
+
+    definition {
+      partition-columns = ["tags:map"]
+      data-columns = [ "timestamp:ts", "min:double", "max:double", "sum:double", "count:double", "avg:double" ]
+      row-key-columns = [ "timestamp" ]
+      downsamplers = [ ]
+    }
+
+    options {
+      shardKeyColumns = [ "__name__", "_ns" ]
+      ignoreShardKeyColumnSuffixes = { "__name__" = ["_bucket", "_count", "_sum"] }
+      valueColumn = "avg"
+      metricColumn = "__name__"
+      ignoreTagsOnPartitionKeyHash = [ "le" ]
+      copyTags = { }
+    }
+
     num-shards = 4
     min-num-nodes = 2
     # Length of chunks to be written, roughly
@@ -29,6 +46,10 @@
 
         max-chunks-size = 400
 
+        # Write buffer size, in bytes, for blob columns (histograms, UTF8Strings).  Since these are variable data types,
+        # we need a maximum size, not a maximum number of items.
+        max-blob-buffer-size = 15000
+
         # Number of bytes of offheap mem to allocate to chunk storage in each shard.  Ex. 1000MB, 1G, 2GB
         # Assume 5 bytes per sample, should be roughly equal to (# samples per time series) * (# time series)
         shard-mem-size = 512MB
@@ -54,5 +75,7 @@
 
         # Amount of time to delay before retrying
         # retry-delay = 15s
+
+        evicted-pk-bloom-filter-capacity = 50000
       }
     }