--- a/conf/timeseries-filodb-server-consul.conf
+++ b/conf/timeseries-filodb-server-consul.conf
@@ -13,6 +13,10 @@ filodb {
     port = 9042
     partition-list-num-groups = 1
   }
+  dataset-configs = [
+    "conf/timeseries-dev-source.conf"
+    "conf/timeseries-ds-1m-dev-source.conf"
+  ]
 }
 
 kamon {