--- a/conf/timeseries-filodb-server.conf
+++ b/conf/timeseries-filodb-server.conf
@@ -5,6 +5,10 @@ filodb {
     port = 9042
     partition-list-num-groups = 1
   }
+  dataset-configs = [
+    "conf/timeseries-dev-source.conf"
+    "conf/timeseries-ds-1m-dev-source.conf"
+  ]
   spread-default = 1
 
   # Override default spread for application using override block which will have non metric shard keys and spread.