--- a/conf/timeseries-standalonetest-source.conf
+++ b/conf/timeseries-standalonetest-source.conf
@@ -1,4 +1,21 @@
     dataset = "prometheus"
+
+    definition {
+      partition-columns = ["tags:map"]
+      data-columns = ["timestamp:ts", "value:double"]
+      row-key-columns = [ "timestamp" ]
+      downsamplers = [ ]
+    }
+
+    options {
+      shardKeyColumns = [ "__name__", "_ns" ]
+      ignoreShardKeyColumnSuffixes = { "__name__" = ["_bucket", "_count", "_sum"] }
+      valueColumn = "value"
+      metricColumn = "__name__"
+      ignoreTagsOnPartitionKeyHash = [ "le" ]
+      copyTags = { }
+    }
+
     num-shards = 4
     min-num-nodes = 2
     sourcefactory = "filodb.kafka.KafkaIngestionStreamFactory"