--- a/coordinator/src/main/scala/filodb.coordinator/NodeCoordinatorActor.scala
+++ b/coordinator/src/main/scala/filodb.coordinator/NodeCoordinatorActor.scala
@@ -8,7 +8,7 @@ import scala.concurrent.duration._
 import akka.actor.{ActorRef, OneForOneStrategy, PoisonPill, Props, Terminated}
 import akka.actor.SupervisorStrategy.{Restart, Stop}
 import akka.event.LoggingReceive
-import com.typesafe.config.Config
+import com.typesafe.config.{Config, ConfigFactory}
 import net.ceedubs.ficus.Ficus._
 
 import filodb.coordinator.client.MiscCommands
@@ -16,7 +16,7 @@ import filodb.core._
 import filodb.core.downsample.DownsampleConfig
 import filodb.core.memstore.MemStore
 import filodb.core.metadata._
-import filodb.core.store.{MetaStore, StoreConfig}
+import filodb.core.store.{IngestionConfig, MetaStore, StoreConfig}
 import filodb.query.QueryCommand
 
 /**
@@ -80,25 +80,31 @@ private[filodb] final class NodeCoordinatorActor(metaStore: MetaStore,
     ingesters.get(dataset).map(func).getOrElse(originator ! UnknownDataset)
   }
 
-  // NOTE: if no QueryActor is set up, forward to NodeGuardian who will then pass it on to a random shard
+  // For now, datasets need to be set up for ingestion before they can be queried (in-mem only)
+  // TODO: if we ever support query API against cold (not in memory) datasets, change this
   private def withQueryActor(originator: ActorRef, dataset: DatasetRef)(func: ActorRef => Unit): Unit =
-    queryActors.get(dataset).map(func).getOrElse {
-      context.parent forward NodeProtocol.ForwardMsg(dataset, func)
-    }
+    queryActors.get(dataset).map(func).getOrElse(originator ! UnknownDataset)
 
+  // Used only for testing
   private def createDataset(originator: ActorRef,
                             datasetObj: Dataset): Unit = {
     (for {
-      resp1 <- metaStore.newDataset(datasetObj) if resp1 == Success
       resp2 <- memStore.store.initialize(datasetObj.ref) if resp2 == Success
     }
     yield {
       originator ! DatasetCreated
     }).recover {
-      case e: NoSuchElementException => originator ! DatasetAlreadyExists
-      case e: StorageEngineException => originator ! e
       case e: Exception => originator ! DatasetError(e.toString)
     }
+  }
+
+  private def initializeDataset(dataset: Dataset, ingestConfig: IngestionConfig): Unit = {
+    logger.info(s"Initializing dataset ${dataset.ref}")
+    memStore.store.initialize(dataset.ref)
+    setupDataset( dataset,
+                  ingestConfig.storeConfig,
+                  IngestionSource(ingestConfig.streamFactoryClass, ingestConfig.sourceConfig),
+                  ingestConfig.downsampleConfig)
   }
 
   // TODO: move createDataset and truncateDataset into NodeClusterActor.  truncate() needs distributed coord
@@ -124,8 +130,7 @@ private[filodb] final class NodeCoordinatorActor(metaStore: MetaStore,
   private def setupDataset(dataset: Dataset,
                            storeConf: StoreConfig,
                            source: IngestionSource,
-                           downsample: DownsampleConfig,
-                           origin: ActorRef): Unit = {
+                           downsample: DownsampleConfig): Unit = {
     import ActorName.{Ingestion, Query}
 
     logger.debug(s"Recreated dataset $dataset from string")
@@ -153,6 +158,7 @@ private[filodb] final class NodeCoordinatorActor(metaStore: MetaStore,
 
   def datasetHandlers: Receive = LoggingReceive {
     case CreateDataset(datasetObj, db) =>
+      // used only for unit testing now
       createDataset(sender(), datasetObj.copy(database = db))
 
     case TruncateDataset(ref) =>
@@ -160,9 +166,9 @@ private[filodb] final class NodeCoordinatorActor(metaStore: MetaStore,
   }
 
   def ingestHandlers: Receive = LoggingReceive {
-    case DatasetSetup(compactDSString, storeConf, source, downsample) =>
-      val dataset = Dataset.fromCompactString(compactDSString)
-      if (!(ingesters contains dataset.ref)) { setupDataset(dataset, storeConf, source, downsample, sender()) }
+    case SetupDataset(dataset, resources, source, storeConf, downsample) =>
+      // used only in unit tests
+      if (!(ingesters contains dataset.ref)) { setupDataset(dataset, storeConf, source, downsample) }
 
     case IngestRows(dataset, shard, rows) =>
       withIngester(sender(), dataset) { _ ! IngestionActor.IngestRows(sender(), shard, rows) }
@@ -205,6 +211,14 @@ private[filodb] final class NodeCoordinatorActor(metaStore: MetaStore,
     } else {
       statusActor.get ! e.clusterActor    // update proxy.  NOTE: this is temporary fix
     }
+
+    settings.datasets.foreach { d =>
+      val config = ConfigFactory.parseFile(new java.io.File(d))
+      val dataset = Dataset.fromConfig(config)
+      val ingestion = IngestionConfig(config, NodeClusterActor.noOpSource.streamFactoryClass).get
+      initializeDataset(dataset, ingestion)
+    }
+
   }
 
   /** Forwards shard actions to the ingester for the given dataset.