--- a/core/src/main/resources/filodb-defaults.conf
+++ b/core/src/main/resources/filodb-defaults.conf
@@ -1,5 +1,8 @@
 filodb {
 
+  # list of paths to dataset + ingestion config files
+  dataset-configs = [ ]
+
   tasks {
     # Frequency with which new shard maps are published
     shardmap-publish-frequency = 5s