--- a/core/src/test/resources/application_test.conf
+++ b/core/src/test/resources/application_test.conf
@@ -1,4 +1,6 @@
 filodb {
+  dataset-configs = [ ]
+
   cassandra {
     hosts = ["localhost"]
     port = 9042