**Pull Request checklist**

- [x] The commit(s) message(s) follows the contribution [guidelines](CONTRIBUTING.md) ?
- [x] Tests for the changes have been added (for bug fixes / features) ?
- [x] Docs have been added / updated (for bug fixes / features) ?

**Current behavior :** (link exiting issues here : https://help.github.com/articles/basic-writing-and-formatting-syntax/#referencing-issues-and-pull-requests)

Require following steps for dataset creation
* initDataset
* createDataset
* setupDataset 

In addition, cluster singleton would send out setup dataset messages on start shard ingestion events.

Third, there was a limitation that spare nodes cannot handle queries since the Query actors are not setup. There is a workaround in place for that, but it is temporary

**New behavior :**

* Remove dataset creation/setup from CLI
* Dataset is now configured via config supplied as part of application.conf
* Ingestion source config also contains the dataset definition. The source path is included in the server config.
* Every NodeCoordinator will setup every dataset from server config on startup.
* IngestionConfig table and DatasetTable have been removed from C* schema.

New Sequence of Events:
* Node Cluster Actor loads all the datasets from this config and initializes ShardManager.
* Node Cluster Actor sends out a CoordinatorRegistered to the registering coodinator.
* Upon receiving that message coordinator initializes all the datasets on the node.
* ShardManager now does not send SetupDataset messages during shard assignment.

Unit Testing:
* SetupDataset case class is now used only for unit testing
* Some unit tests need to initialize both NodeCoodinatorActor and NodeClusterActor with the setup message if the dataset config is not present in the server config.

**BREAKING CHANGES**

* Significant changes in this PR.
* Need config changes for deployment
* Cannot do rolling upgrades

Benchmarks are still working:


Before
```
[info] Benchmark                                                        Mode  Cnt      Score     Error  Units
[info] QueryAndIngestBenchmark.parallelQueries                         thrpt   45    298.498 ±  17.089  ops/s
[info] QueryAndIngestBenchmark.parallelQueries:JFR                     thrpt             NaN              N/A
[info] QueryInMemoryBenchmark.noOverlapQueries                         thrpt   45   9470.662 ± 200.798  ops/s
[info] QueryInMemoryBenchmark.noOverlapQueries:JFR                     thrpt             NaN              N/A
[info] QueryInMemoryBenchmark.singleThreadedMinOverTimeQuery           thrpt   45   9290.641 ±  97.687  ops/s
[info] QueryInMemoryBenchmark.singleThreadedMinOverTimeQuery:JFR       thrpt             NaN              N/A
[info] QueryInMemoryBenchmark.singleThreadedRawQuery                   thrpt   45  12795.824 ± 246.735  ops/s
[info] QueryInMemoryBenchmark.singleThreadedRawQuery:JFR               thrpt             NaN              N/A
[info] QueryInMemoryBenchmark.singleThreadedSumRateCCQuery             thrpt   45   3896.019 ± 109.958  ops/s
[info] QueryInMemoryBenchmark.singleThreadedSumRateCCQuery:JFR         thrpt             NaN              N/A
[info] QueryInMemoryBenchmark.someOverlapQueries                       thrpt   45   8372.649 ± 192.382  ops/s
[info] QueryInMemoryBenchmark.someOverlapQueries:JFR                   thrpt             NaN              N/A
[info] QueryOnDemandBenchmark.parallelQueries                          thrpt   45    683.805 ± 114.539  ops/s
[info] QueryOnDemandBenchmark.parallelQueries:JFR                      thrpt             NaN              N/A
[info] QueryHiCardInMemoryBenchmark.scanSumOfRateBenchmark              avgt   45  15968.221 ± 926.313  us/op
[info] QueryHiCardInMemoryBenchmark.scanSumOfRateBenchmark:JFR          avgt             NaN              N/A
[info] QueryHiCardInMemoryBenchmark.scanSumOfSumOverTimeBenchmark       avgt   45   1953.259 ± 294.274  us/op
[info] QueryHiCardInMemoryBenchmark.scanSumOfSumOverTimeBenchmark:JFR   avgt             NaN              N/A
[info]
```

After

```
[info] Benchmark                                                        Mode  Cnt      Score      Error  Units
[info] QueryAndIngestBenchmark.parallelQueries                         thrpt   45    289.074 ±   16.577  ops/s
[info] QueryAndIngestBenchmark.parallelQueries:JFR                     thrpt             NaN               N/A
[info] QueryInMemoryBenchmark.noOverlapQueries                         thrpt   45   8653.319 ±  803.506  ops/s
[info] QueryInMemoryBenchmark.noOverlapQueries:JFR                     thrpt             NaN               N/A
[info] QueryInMemoryBenchmark.singleThreadedMinOverTimeQuery           thrpt   45   9000.817 ±  369.030  ops/s
[info] QueryInMemoryBenchmark.singleThreadedMinOverTimeQuery:JFR       thrpt             NaN               N/A
[info] QueryInMemoryBenchmark.singleThreadedRawQuery                   thrpt   45  12921.314 ±  221.268  ops/s
[info] QueryInMemoryBenchmark.singleThreadedRawQuery:JFR               thrpt             NaN               N/A
[info] QueryInMemoryBenchmark.singleThreadedSumRateCCQuery             thrpt   45   4136.640 ±   23.669  ops/s
[info] QueryInMemoryBenchmark.singleThreadedSumRateCCQuery:JFR         thrpt             NaN               N/A
[info] QueryInMemoryBenchmark.someOverlapQueries                       thrpt   45   8769.869 ±  117.805  ops/s
[info] QueryInMemoryBenchmark.someOverlapQueries:JFR                   thrpt             NaN               N/A
[info] QueryOnDemandBenchmark.parallelQueries                          thrpt   45    949.192 ±   51.742  ops/s
[info] QueryOnDemandBenchmark.parallelQueries:JFR                      thrpt             NaN               N/A
[info] QueryHiCardInMemoryBenchmark.scanSumOfRateBenchmark              avgt   45  15807.801 ± 1566.297  us/op
[info] QueryHiCardInMemoryBenchmark.scanSumOfRateBenchmark:JFR          avgt             NaN               N/A
[info] QueryHiCardInMemoryBenchmark.scanSumOfSumOverTimeBenchmark       avgt   45   1887.235 ±   90.744  us/op
[info] QueryHiCardInMemoryBenchmark.scanSumOfSumOverTimeBenchmark:JFR   avgt             NaN               N/A


```