@velvia 2019-05-01
Oh, I would say that Coordinators also should not wait for the CoordinatorRegistered message.  They should just initialize themselves at startup.