@vishramachandran 2019-05-01
> Oh, I would say that Coordinators also should not wait for the CoordinatorRegistered message. They should just initialize themselves at startup.

There are race conditions with doing it that way. For one, the cluster Singleton may not have been initialized. Remember that statusActor needs reference to the clusterSingleton. Hence this approach is necessary.