@velvia 2019-05-02
@vishramachandran  hmmm.  I wonder if we should just change the NodeCoordinator to take in an ActorRef to the cluster actor, ie make sure ClusterActor is initialized first.   That would simplify a lot of things including the whole workflow.  The only reason why things are done the current way is that in the CLI, the NodeCoordinator is started up without the ClusterActor, but that was for a use case (CSV ingestion and export) which we aren't going to support anymore.