@vishramachandran 2019-05-06
> I wonder if we should just change the NodeCoordinator to take in an ActorRef to the cluster actor, ie make sure ClusterActor is initialized first.

@velvia The cyclic dependency needs to be broken. Shard Manager on cluster singleton depends on NodeCoord registration. NodeCoord cannot wait for singleton to have started. 

Moreover, we have to address the case where singleton and NodeCoord are on different nodes.